added ServSpecOpt CarryingCapacityMod (double default 1.0) as * modifier for mobile max_weight changed armorzone sets to unsigned short

+++ pol-core/pol/item/armrtmpl.h
@@ -28,7 +28,7 @@ public:
 	virtual void PopulateStruct( BStruct* descriptor ) const;
 
     unsigned short ar;
-    std::set<unsigned> zones;
+    std::set<unsigned short> zones;
     ScriptDef on_hit_script;
 };
